pyproject.toml
[project]
name = "scrapytut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "botocore>=1.42.71",
    "mysql-connector-python>=9.6.0",
    "playwright>=1.58.0",
    "psycopg2>=2.9.11",
    "scrapeops-scrapy-proxy-sdk>=1.0",
    "scrapy-playwright>=0.0.46",
    "scrapy-selenium>=0.0.7",
    "scrapy-splash>=0.11.1",
    "scrapy-user-agents>=0.1.1",
    "selenium==4.10",
]
